tirando a logica do componente botao-palpite
